Add debuginfod stack symbolization
Add a debuginfod symbolizer mode that normalizes BPF frames, maps them to module build IDs, fetches debuginfo through the upstream debuginfod client, and resolves symbols with blazesym.

Keep the existing HTTP symbolizer path for batch services, add a mock debuginfod test that exercises fetch, cache, and symbolization, and document the DEBUGINFOD_URLS/DEBUGINFOD_CACHE_PATH flow.

Also record the clippy workflow in AGENTS.md and apply the small clippy cleanups found while validating this change.

Co-developed-by: Codex (GPT-5) <noreply@openai.com>

+++ crates/corefile/src/lib.rs
@@ -228,7 +228,7 @@ fn statvfs(path: &Path) -> Result<FsStats> {
         return Err(std::io::Error::last_os_error()).context("statvfs output dir");
     }
     let stat = unsafe { stat.assume_init() };
-    let block_size = stat.f_frsize.max(stat.f_bsize) as u64;
+    let block_size = stat.f_frsize.max(stat.f_bsize);
     Ok(FsStats {
         total_bytes: block_size.saturating_mul(stat.f_blocks),
         available_bytes: block_size.saturating_mul(stat.f_bavail),
